Expose HybridAStarConfig and MpcConfig to TypeScript via wasm_bindgen; remove CAR_CONSTANTS and migrate algorithm constants to Rust
Rust:
- HybridAStarConfig: add #[wasm_bindgen] with constructor, all getters, and step_budget field (replaces TS HYBRID_STEP_BUDGET)
- MpcConfig: add #[wasm_bindgen] with constructor, all getters, and dt field (replaces TS DEFAULT_LOCAL_PLANNER_DT)
- horizon_length/max_iter/num_steer_commands/step_budget use u32 for wasm compat
- Merge mpc_control_preview + mpc_control_preview_with_config into one function taking &MpcConfig (config.dt replaces dt param)
- Remove dt parameter from linear_mpc_control and build_mpc_problem (use config.dt)
- Export HybridAStarConfig and MpcConfig from lib.rs
- Disable wasm-opt in wasm-pack (unavailable in CI environment)

TypeScript:
- Remove CAR_CONSTANTS from appModel.ts (values live in CarConfig in Rust)
- Remove DEFAULT_CAR_SHAPE and DEFAULT_MOTION_LIMITS from appTypes.ts
- App.tsx: carShape/motionLimits state nullable (set after WASM loads); MapServerNode init moved to WASM useEffect; remove LOCAL_PLANNER_DT
- MapViewport + mapViewportDraw: accept nullable carShape/motionLimits
- workerTypes.ts: remove HYBRID_STEP_BUDGET and DEFAULT_LOCAL_PLANNER_DT; remove simDeltaTime from LocalPlannerSession
- workerHandlers.ts: use new HybridAStarConfig().step_budget
- workerCodecs.ts: create MpcConfig per runLocalPlannerUpdate call; pass to mpc_control_preview
- workerHelpers.ts/computeWorker.ts/wasmCore.ts/useSimulationSetup.ts: remove dt param threading

Agent-Logs-Url: https://github.com/FredBill1/AutonomousDrivingDemoWASM/sessions/a568fa23-d6f1-4d3c-9afd-d920005c0889

+++ wasm-core/src/hybrid_astar/planner.rs
@@ -302,7 +302,7 @@ impl HybridAStarPlanner {
         nodes.insert(start_ijk, start_node);
 
         Ok(Self {
-            steer_commands: steer_commands(car_config.target_max_steer(), ha_config.num_steer_commands),
+            steer_commands: steer_commands(car_config.target_max_steer(), ha_config.num_steer_commands as usize),
             car_config,
             ha_config,
             heuristic,
